+++ src/components/tyhUi/packages/link/TyhLink.vue
@@ -51,7 +51,9 @@ const isClass = computed(() => {
   return [
     'tyh-link',
     `tyh-link-${props.type}`,
-    props.prohibit ? `tyh-link-prohibit-${props.type}` : ''
+    {
+      [`tyh-link-prohibit-${props.type}`]: props.prohibit
+    }
   ]
 })
 </script>