+++ src/components/tyhUi/packages/tagging/TyhTagging.vue
@@ -1,5 +1,5 @@
 <template>
-  <span :class="isClass">
+  <span :class="isClass" :style="isStyle">
     <slot />
   </span>
 </template>
@@ -13,14 +13,20 @@ const props = defineProps({
     validator (val) {
       return ['default', 'primary', 'success', 'danger', 'warning'].includes(val)
     }
+  },
+  size: {
+    type: [String, Number],
+    default: '13'
+  },
+  color: {
+    type: String,
+    default: '#333'
   }
 })
-const isClass = () => {
-  return [
-    'tyh-tagging',
-    `tyh-tagging-${props.type}`
-  ]
-}
+const isClass = computed(() => ['tyh-tagging', `tyh-tagging-${props.type}`])
+const isStyle = computed(() => {
+  return { fontSize: `${props.size}px`, color: props.color }
+})
 </script>
 
 <style src="./style/index.css" scoped>