+++ src/components/tyhUi/packages/text/TyhText.vue
@@ -30,7 +30,7 @@ defineProps({
     }
   },
   size: {
-    type: String,
+    type: [String, Number],
     default: '16'
   },
   block: Boolean,