+++ src/components/tyhUi/src/index.js
@@ -85,7 +85,7 @@ const install = app => {
 }
 
 const tyhUi = {
-  version: "3.2.9",
+  version: "3.2.10",
   install
 }